fix: fixes done as per the review comments
