+++ src/data/services/lms/urls.js
@@ -30,6 +30,8 @@ export const creditPurchaseUrl = (courseId) => {
 };
 export const creditRequestUrl = (providerId) => `${getApiUrl()}/credit/v1/providers/${providerId}/request/`;
 
+export const programsApiUrl = () => `${getApiUrl()}/dashboard/v0/programs/`;
+
 export default StrictDict({
   getApiUrl,
   baseAppUrl,
@@ -42,4 +44,5 @@ export default StrictDict({
   learningMfeUrl,
   programsUrl,
   updateEmailSettings,
+  programsApiUrl,
 });
