--- src/data/types/index.ts
@@ -1,8 +1,0 @@
-import { ProgramData } from '../../containers/ProgramDashboard/data/types';
-
-export interface ProgramsListData {
-  data: ProgramData[];
-  isLoading: boolean;
-  isError: boolean;
-  error: Error | null;
-}
