+++ src/index.jsx
@@ -53,7 +53,7 @@ subscribe(APP_READY, () => {
             <Routes>
               <Route path="/" element={<PageWrap><App /></PageWrap>} />
               {getConfig().ENABLE_PROGRAM_DASHBOARD && (
-                <Route path="programs" element={<PageWrap><ProgramsList /></PageWrap>} />
+                <Route path="/programs" element={<PageWrap><ProgramsList /></PageWrap>} />
               )}
               <Route path="*" element={<Navigate to="/" replace />} />
             </Routes>
